requirements/quality.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/quality.txt requirements/base.in requirements/test-master.in requirements/dev.in requirements/quality.in requirements/test.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
argparse==1.4.0           # via caniusepython3
asn1crypto==0.24.0        # via cryptography
astroid==1.5.2            # via edx-lint, pylint, pylint-celery, pylint-plugin-utils
attrs==17.4.0             # via pytest
awscli==1.11.178
backports.functools-lru-cache==1.5  # via astroid, caniusepython3, pylint
bcrypt==3.1.4             # via paramiko
billiard==3.3.0.23        # via celery
boto3==1.4.7
botocore==1.7.36          # via awscli, boto3, s3transfer
caniusepython3==6.0.0
celery==3.1.18
cffi==1.11.4              # via bcrypt, cryptography, pynacl
click-log==0.1.8          # via edx-lint
click==6.7                # via click-log, edx-lint, pip-tools
colorama==0.3.7           # via awscli
configparser==3.5.0       # via pydocstyle, pylint
cookies==2.2.1            # via responses
coverage==4.5.1           # via pytest-cov
cryptography==1.9
ddt==1.1.1
distlib==0.2.6            # via caniusepython3
django-fernet-fields==0.5
django-model-utils==3.1.1
django==1.11.5
djangorestframework-jwt==1.11.0  # via edx-drf-extensions
djangorestframework==3.6.2
docutils==0.14            # via awscli, botocore
edx-drf-extensions==1.2.3
edx-lint==0.5.5
edx-rest-api-client==1.7.1
enum34==1.1.6             # via astroid, cryptography
factory-boy==2.10.0
faker==0.8.11             # via factory-boy
first==2.0.1              # via pip-tools
flaky==3.4.0
freezegun==0.3.9
funcsigs==1.0.2           # via mock, pytest
future==0.16.0            # via vertica-python
futures==3.2.0 ; python_version == "2.7"
idna==2.6                 # via cryptography
ipaddress==1.0.19         # via cryptography, faker
isort==4.3.4
jmespath==0.9.3           # via boto3, botocore
kombu==3.0.37             # via celery
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
mock==2.0.0
mysql-connector==2.2.3
packaging==16.8           # via caniusepython3
paramiko==2.4
pbr==3.1.1                # via mock
pip-tools==1.11.0
pluggy==0.6.0             # via pytest, tox
py==1.5.2                 # via pytest, pytest-catchlog, tox
pyasn1==0.4.2             # via paramiko, rsa
pycodestyle==2.3.1
pycparser==2.18           # via cffi
pydocstyle==2.1.1
pyjwt==1.5.3              # via djangorestframework-jwt, edx-rest-api-client
pylint-celery==0.3        # via edx-lint
pylint-django==0.7.2      # via edx-lint
pylint-plugin-utils==0.2.6  # via pylint-celery, pylint-django
pylint==1.7.1             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyminizip==0.2.1
pynacl==1.2.1             # via paramiko
pyparsing==2.2.0          # via packaging
pytest-catchlog==1.2.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.4.0             # via pytest-catchlog, pytest-cov, pytest-django
python-dateutil==2.6.1    # via botocore, edx-drf-extensions, faker, freezegun, vertica-python
pytz==2018.3              # via celery, django, vertica-python
pyyaml==3.12              # via awscli
requests==2.9.1
responses==0.8.1
rsa==3.4.2                # via awscli
s3transfer==0.1.12        # via awscli, boto3
singledispatch==3.4.0.3   # via astroid, pylint
six==1.11.0               # via astroid, bcrypt, cryptography, edx-lint, faker, freezegun, mock, packaging, pip-tools, pydocstyle, pylint, pynacl, pytest, python-dateutil, responses, singledispatch, tox, vertica-python
slumber==0.7.1            # via edx-rest-api-client
snowballstemmer==1.2.1    # via pydocstyle
testfixtures==5.4.0
text-unidecode==1.1       # via faker
tox==2.9.1
unicodecsv==0.14.1
vertica-python==0.7.3
virtualenv==15.1.0        # via tox
wrapt==1.10.11            # via astroid
